Button clicks reported to gameboard
